Convert project to Maven
Restructure project directories to conform with Maven standard.
Resources got moved, that resulted in minor code change.
There are items left to complete that need discussion, most importantly, the license as it will affect where the artifact will reside.
I still have an outstanding question about how the release will work.
This code generates web pages to host the jar on GitHub. Some GitHub work needs to be done to support this.
