+++ java.header
@@ -1,0 +1,3 @@
+^/\*$
+^ \* Copyright \(c\) \d\d\d\d\. The League of Amazing Programmers\. All Rights Reserved\.$
+^ \*/$
