+++ robot-examples/src/main/java/org/jointheleague/robot/example/ExtendedKeyboardAdapter.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2016. The League of Amazing Programmers. All Rights Reserved.
+ */
+
+package org.jointheleague.robot.example;
+
+import org.jointheleague.graphical.robot.KeyboardAdapter;
+
+import java.awt.event.KeyEvent;
+
+public class ExtendedKeyboardAdapter extends KeyboardAdapter {
+
+
+    public ExtendedKeyboardAdapter() {
+        super();
+    }
+
+
+    @Override
+    public void keyTyped(KeyEvent e) {
+        switch (e.getKeyChar()) {
+            case 'm':
+                robot.miniaturize();
+                break;
+            case 'M':
+                robot.expand();
+                break;
+            case 'u':
+                robot.penUp();
+                break;
+            case 'd':
+                robot.penDown();
+                break;
+            default:
+                super.keyTyped(e);
+        }
+
+    }
+
+}
