+++ robot-examples/src/main/java/org/jointheleague/robot/example/RobotExample6.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2016. The League of Amazing Programmers. All Rights Reserved.
+ */
+
+package org.jointheleague.robot.example;
+
+import org.jointheleague.graphical.robot.Robot;
+
+public class RobotExample6 {
+
+    public static void main(String[] args) throws InterruptedException {
+        Robot vic = new Robot("vic", 600, 300);
+        Robot june = new Robot("june", 300, 300);
+        june.setSpeed(10);
+        vic.setSpeed(10);
+        vic.addKeyboardAdapter(new ShiftKeyboardAdapter(true));
+        june.addKeyboardAdapter(new ShiftKeyboardAdapter(false));
+    }
+}
