+++ robot/src/main/resources/batman.robi
