+++ robot/src/main/resources/formersecretaryofstatecondoleezzarice.robi
