+++ robot/src/main/resources/league_logo.png
