+++ robot/src/main/resources/mini.robi
