+++ robot/src/main/resources/vic.robi
